requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cloudinary==1.37.0
crispy-bootstrap5==2023.10
distlib==0.3.8
Django==5.0.1
django-crispy-forms==2.1
environ==1.0
factory-boy==3.3.0
Faker==22.1.0
filelock==3.13.1
gunicorn==21.2.0
packaging==23.2
Pillow==10.1.0
pipenv==2023.11.15
platformdirs==4.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.25.0
